Changes namespace & updates classes/comments/tests accordingly
* Updates namespace from com\google\i18n\phonenumbers to libphonenumber
* Updates comments & README for namespace change
* Moves multiple classes in single file to individual classes
* Adds use & require_once statements in tests, build & demo.php accordingly
