+++ PhoneNumber.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace com\google\i18n\phonenumbers;
+namespace libphonenumber;
 
 class PhoneNumber {
 
@@ -188,12 +188,4 @@ class PhoneNumber {
 		}
 		return $this;
 	}
-}
-class CountryCodeSource {
-
-	const FROM_NUMBER_WITH_PLUS_SIGN = 0;
-	const FROM_NUMBER_WITH_IDD = 1;
-	const FROM_NUMBER_WITHOUT_PLUS_SIGN = 2;
-	const FROM_DEFAULT_COUNTRY = 3;
-
 }
