+++ demo.php
@@ -1,8 +1,16 @@
 <?php
-use com\google\i18n\phonenumbers\PhoneNumberUtil;
-use com\google\i18n\phonenumbers\PhoneNumber;
+namespace libphonenumber;
 
-require_once 'PhoneNumberUtil.php';
+require_once dirname(__FILE__) . '/PhoneNumberUtil.php';
+require_once dirname(__FILE__) . '/CountryCodeToRegionCodeMap.php';
+require_once dirname(__FILE__) . '/PhoneNumber.php';
+require_once dirname(__FILE__) . '/PhoneMetadata.php';
+require_once dirname(__FILE__) . '/PhoneNumberDesc.php';
+require_once dirname(__FILE__) . '/NumberFormat.php';
+require_once dirname(__FILE__) . '/Matcher.php';
+require_once dirname(__FILE__) . '/CountryCodeSource.php';
+require_once dirname(__FILE__) . '/PhoneNumberType.php';
+require_once dirname(__FILE__) . '/PhoneNumberFormat.php';
 
 $swissNumberStr = "044 668 18 00";
 $phoneUtil = PhoneNumberUtil::getInstance();
